[IMP] account,*: move accrued orders to stock modules
* - purchase
  - purchase_stock
  - sale
  - sale_stock
  - stock_account

Accrued orders in accounting use some code from the stock module.
Considering that account does not depend on stock, this functionality
should be moved to related stock modules (account -> stock_account,
purchase -> purchase_stock, sale -> sale_stock).

[IMP] sale_timesheet: run accrued order tests if sale_stock is installed

Accrued orders have been moved to sale_stock from sale, which means
that the tests for this model should be run only if sale_stock is
installed. This check is added since it is better that creating
a separate module just to run the tests conditionally.

task-6395492
