+++ addons/account/security/ir.access.csv
@@ -102,7 +102,6 @@ access_account_financial_year_op,access.account.financial.year.op,account.financ
 access_account_setup_bank_manual_config,access.account.setup.bank.manual.config,account.setup.bank.manual.config,account.group_account_manager,cru,
 account_move_send_single_rule_group_invoice,Readonly Invoice Send and Print (single),account.move.send.wizard,account.group_account_invoice,crud,
 account_move_send_batch_rule_group_invoice,Readonly Invoice Send and Print (batch),account.move.send.batch.wizard,account.group_account_invoice,crud,
-access_account_accrued_orders_wizard,account.account.accrued.orders.wizard,account.accrued.orders.wizard,account.group_account_user,cru,
 access_account_report_readonly,account.report.readonly,account.report,account.group_account_readonly,r,
 access_account_report_ac_user,account.report.ac.user,account.report,account.group_account_manager,crud,
 access_account_report_line_readonly,account.report.line.readonly,account.report.line,account.group_account_readonly,r,
