+++ addons/account/wizard/__init__.py
@@ -12,7 +12,6 @@ from . import account_move_send_wizard
 from . import account_move_send_batch_wizard
 from . import base_document_layout
 from . import account_payment_register
-from . import accrued_orders
 from . import base_partner_merge
 from . import account_merge_wizard
 from . import account_mail_compose_message_wizard
