+++ addons/purchase/tests/__init__.py
@@ -8,7 +8,6 @@ from . import test_purchase_order_product_catalog
 from . import test_purchase_order_report
 from . import test_purchase_invoice
 from . import test_access_rights
-from . import test_accrued_purchase_orders
 from . import test_purchase_dashboard
 from . import test_import_files
 from . import test_purchase_product_catalog
