+++ addons/purchase/views/purchase_views.xml
@@ -1001,15 +1001,6 @@
         </field>
     </record>
 
-    <record id="action_accrued_expense_entry" model="ir.actions.act_window">
-        <field name="name">Accrued Expense Entry</field>
-        <field name="res_model">account.accrued.orders.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="binding_model_id" ref="purchase.model_purchase_order"/>
-        <field name="group_ids" eval="[(4, ref('account.group_account_user'))]"/>
-        <field name="target">new</field>
-    </record>
-
     <record id="action_merger" model="ir.actions.server">
         <field name="name">Merge RFQs</field>
         <field name="model_id" ref="purchase.model_purchase_order"/>
