+++ addons/purchase_stock/views/purchase_views.xml
@@ -101,4 +101,13 @@
             </xpath>
         </field>
     </record>
+
+    <record id="action_accrued_expense_entry" model="ir.actions.act_window">
+        <field name="name">Accrued Expense Entry</field>
+        <field name="res_model">stock_account.accrued.orders.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="binding_model_id" ref="purchase.model_purchase_order"/>
+        <field name="group_ids" eval="[(4, ref('account.group_account_user'))]"/>
+        <field name="target">new</field>
+    </record>
 </odoo>
