+++ addons/sale/tests/__init__.py
@@ -2,7 +2,6 @@
 
 from . import (
     test_access_rights,
-    test_accrued_sale_orders,
     test_analytic_to_sale_to_invoice,
     test_common,
     test_controllers,
