+++ addons/sale_stock/tests/__init__.py
@@ -16,3 +16,4 @@ from . import test_sale_stock_accrued_entries
 from . import test_sale_stock_access_rights
 from . import test_create_perf
 from . import test_packaging_tours
+from . import test_accrued_sale_orders
