+++ addons/sale_stock/wizard/stock_account_accrued_orders_wizard_views.xml
@@ -3,7 +3,7 @@
 
     <record id="action_accrued_revenue_entry" model="ir.actions.act_window">
         <field name="name">Accrued Revenue Entry</field>
-        <field name="res_model">account.accrued.orders.wizard</field>
+        <field name="res_model">stock_account.accrued.orders.wizard</field>
         <field name="view_mode">form</field>
         <field name="binding_model_id" ref="sale.model_sale_order"/>
         <field name="group_ids" eval="[(4, ref('account.group_account_user'))]"/>
@@ -12,7 +12,7 @@
 
     <record id="action_accrued_revenue_entry_sale_order_line" model="ir.actions.act_window">
         <field name="name">Accrued Revenue Entry</field>
-        <field name="res_model">account.accrued.orders.wizard</field>
+        <field name="res_model">stock_account.accrued.orders.wizard</field>
         <field name="view_mode">form</field>
         <field name="binding_model_id" ref="sale.model_sale_order_line"/>
         <field name="group_ids" eval="[(4, ref('account.group_account_user'))]"/>
