+++ addons/stock_account/__manifest__.py
@@ -34,6 +34,7 @@ Dashboard / Reports for Warehouse Management includes:
         'views/stock_lot_views.xml',
         'views/stock_move_views.xml',
         'wizard/stock_inventory_adjustment_name_views.xml',
+        'wizard/accrued_orders.xml',
         'report/account_invoice_report_view.xml',
         'report/stock_avco_audit_report_views.xml',
         'report/stock_valuation_report.xml',
