+++ addons/stock_account/security/ir.access.csv
@@ -9,4 +9,5 @@ access_stock_move_invoicing_payments_readonly,stock.move,stock.move,account.grou
 access_stock_move_invoicing_payments,stock.move,stock.move,account.group_account_invoice,cru,
 access_stock_avco_report_account_readonly,stock.avco.report,stock.avco.report,account.group_account_readonly,r,
 access_stock_avco_report_stock_manager,stock.avco.report,stock.avco.report,stock.group_stock_manager,r,
+access_account_accrued_orders_wizard,stock_account.stock_account.accrued.orders.wizard,stock_account.accrued.orders.wizard,account.group_account_user,cru,
 stock_avco_report_rule,Stock Average Cost Report multi-company,stock.avco.report,,crud,"[('company_id','in', company_ids)]"
