+++ addons/stock_account/wizard/__init__.py
@@ -1,3 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import stock_inventory_adjustment_name
+from . import accrued_orders
