+++ addons/stock_account/wizard/accrued_orders.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="view_account_accrued_orders_wizard" model="ir.ui.view">
-        <field name="name">account.accrued.orders.wizard.view</field>
-        <field name="model">account.accrued.orders.wizard</field>
+        <field name="name">stock_account.accrued.orders.wizard.view</field>
+        <field name="model">stock_account.accrued.orders.wizard</field>
         <field name="arch" type="xml">
             <form string="Make Accrual Entries">
                 <field name="company_id" invisible="1"/>
